requirements.txt
google==3.0.0
fastapi==0.88.0
aspose.slides==22.11.0
psycopg2==2.9.5
sqlalchemy==1.4.44
python-pptx==0.6.21
google-auth-oauthlib==0.7.1
google-api-python-client==2.69.0
uvicorn==0.17.6
starlette==0.22.0